requirements.txt
asgiref==3.7.2
Django==4.2.7
django-environ==0.11.2
django-phonenumber-field==7.2.0
djangorestframework==3.14.0
env==0.1.0
environ==1.0
phonenumbers==8.13.25
psycopg==3.1.12
psycopg-binary==3.1.12
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
